requirements.txt
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
audioread==3.0.0
backcall==0.2.0
blinker==1.6.2
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
contourpy==1.0.7
coverage==7.1.0
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
exceptiongroup==1.1.0
executing==1.2.0
ffmpeg==1.4
filelock==3.12.0
Flask==2.3.2
fonttools==4.39.3
graphics.py==5.0.1.post1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
install==1.3.5
ipykernel==6.21.1
ipytest==0.13.1
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.2
jupyter_core==5.2.0
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.40.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
mutagen==1.46.0
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.0
numpy==1.24.1
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.6.2
pluggy==1.0.0
pooch==1.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.17
Pygments==2.14.0
pyparsing==3.0.9
pytest==7.2.1
pytest-cov==4.0.0
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.0
regex==2023.5.5
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
stack-data==0.6.2
stanfordkarel==0.2.7
sympy==1.11.1
threadpoolctl==3.1.0
tiktoken==0.3.3
tomli==2.0.1
torch==2.0.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==2.0.2
wcwidth==0.2.6
websockets==11.0
Werkzeug==2.3.3
yt-dlp==2023.3.4
zipp==3.15.0
